+++ src/app/api/cron/wiki/scraper/relicfragments/compare.ts
@@ -1,0 +1,19 @@
+import { RelicFragmentItem } from '@/app/(data)/items/types/RelicFragmentItem'
+
+export function relicfragmentDataCompare(
+  newData: {
+    description: string
+  },
+  currentItem: RelicFragmentItem,
+): {
+  descriptionMatches: boolean
+  dataDiffers: boolean
+} {
+  const descriptionMatches = newData.description === currentItem.description
+  const dataDiffers = !descriptionMatches
+
+  return {
+    descriptionMatches,
+    dataDiffers,
+  }
+}