+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Fixed
 
 - Fixed all discrepancies with Mutator item descriptions and stats from the wiki.
+- Fixed all discrepancies with Perk item descriptions and stats from the wiki.
+- Fixed all discrepancies with the Relic Fragment item descriptions and stats from the wiki.
 
 ## 2024-05-04